# docs/frontend/openapi-spec.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "InsightLogger.Api",
    "version": "1.0"
  },
  "paths": {
    "/analyze/build-log": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Analyze a full build or tool log.",
        "operationId": "AnalyzeBuildLog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeBuildLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeBuildLogResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/analyze/compiler-error": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Analyze a single compiler/runtime diagnostic or compact error block.",
        "operationId": "AnalyzeCompilerError",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeCompilerErrorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeCompilerErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/narratives": {
      "get": {
        "tags": ["Analysis History"],
        "summary": "List persisted build-log narratives from prior analyses.",
        "operationId": "GetAnalysisNarratives",
        "parameters": [
          {
            "name": "tool",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "text",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnalysisNarrativesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysisId}": {
      "get": {
        "tags": ["Analysis History"],
        "summary": "Get a full persisted analysis by analysis id.",
        "operationId": "GetAnalysis",
        "parameters": [
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysisId}/narrative": {
      "get": {
        "tags": ["Analysis History"],
        "summary": "Get a persisted grouped diagnostic narrative by analysis id.",
        "operationId": "GetAnalysisNarrative",
        "parameters": [
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnalysisNarrativeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisSummaryContract": {
        "type": "object",
        "properties": {
          "totalDiagnostics": {
            "type": "integer",
            "format": "int32"
          },
          "groupCount": {
            "type": "integer",
            "format": "int32"
          },
          "primaryIssueCount": {
            "type": "integer",
            "format": "int32"
          },
          "errorCount": {
            "type": "integer",
            "format": "int32"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "AnalyzeBuildLogRequest": {
        "type": "object",
        "properties": {
          "tool": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "projectName": {
            "type": "string",
            "nullable": true
          },
          "repository": {
            "type": "string",
            "nullable": true
          },
          "branch": {
            "type": "string",
            "nullable": true
          },
          "commitSha": {
            "type": "string",
            "nullable": true
          },
          "environment": {
            "$ref": "#/components/schemas/AnalyzeEnvironmentContract"
          },
          "options": {
            "$ref": "#/components/schemas/AnalyzeRequestOptionsContract"
          }
        },
        "additionalProperties": false
      },
      "AnalyzeBuildLogResponse": {
        "type": "object",
        "properties": {
          "analysisId": {
            "type": "string",
            "nullable": true
          },
          "toolDetected": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "$ref": "#/components/schemas/AnalysisSummaryContract"
          },
          "rootCauseCandidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RootCauseCandidateContract"
            },
            "nullable": true
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticGroupContract"
            },
            "nullable": true
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticContract"
            },
            "nullable": true
          },
          "matchedRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchedRuleContract"
            },
            "nullable": true
          },
          "narrative": {
            "$ref": "#/components/schemas/AnalysisNarrativeContract"
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingMetadataContract"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AnalyzeCompilerErrorRequest": {
        "type": "object",
        "properties": {
          "tool": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "context": {
            "$ref": "#/components/schemas/AnalyzeRequestContextContract"
          },
          "options": {
            "$ref": "#/components/schemas/AnalyzeRequestOptionsContract"
          }
        },
        "additionalProperties": false
      },
      "AnalyzeCompilerErrorResponse": {
        "type": "object",
        "properties": {
          "fingerprint": {
            "type": "string",
            "nullable": true
          },
          "toolDetected": {
            "type": "string",
            "nullable": true
          },
          "diagnostic": {
            "$ref": "#/components/schemas/DiagnosticContract"
          },
          "explanation": {
            "type": "string",
            "nullable": true
          },
          "likelyCauses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "suggestedFixes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "matchedRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchedRuleContract"
            },
            "nullable": true
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingMetadataContract"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AnalyzeEnvironmentContract": {
        "type": "object",
        "properties": {
          "os": {
            "type": "string",
            "nullable": true
          },
          "ci": {
            "type": "boolean",
            "nullable": true
          },
          "machineName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AnalyzeRequestOptionsContract": {
        "type": "object",
        "properties": {
          "persist": {
            "type": "boolean"
          },
          "useAiEnrichment": {
            "type": "boolean"
          },
          "includeRawDiagnostics": {
            "type": "boolean"
          },
          "includeGroups": {
            "type": "boolean"
          },
          "includeProcessingMetadata": {
            "type": "boolean"
          },
          "useAiRootCauseNarrative": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "additionalProperties": false
      },
...
      "DiagnosticGroupContract": {
        "type": "object",
        "properties": {
          "fingerprint": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "groupReason": {
            "type": "string",
            "nullable": true
          },
          "primaryDiagnosticId": {
            "type": "string",
            "nullable": true
          },
          "relatedDiagnosticIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetAnalysisNarrativeResponse": {
        "type": "object",
        "properties": {
          "analysisId": {
            "type": "string",
            "nullable": true
          },
          "inputType": {
            "type": "string",
            "nullable": true
          },
          "toolDetected": {
            "type": "string",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "$ref": "#/components/schemas/AnalysisSummaryContract"
          },
          "narrative": {
            "$ref": "#/components/schemas/AnalysisNarrativeContract"
          },
          "projectName": {
            "type": "string",
            "nullable": true
          },
          "repository": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetAnalysisNarrativesResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisNarrativeHistoryItemContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetAnalysisResponse": {
        "type": "object",
        "properties": {
          "analysisId": {
            "type": "string",
            "nullable": true
          },
          "inputType": {
            "type": "string",
            "nullable": true
          },
          "toolDetected": {
            "type": "string",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "$ref": "#/components/schemas/AnalysisSummaryContract"
          },
          "rootCauseCandidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RootCauseCandidateContract"
            },
            "nullable": true
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticGroupContract"
            },
            "nullable": true
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticContract"
            },
            "nullable": true
          },
          "matchedRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchedRuleContract"
            },
            "nullable": true
          },
          "narrative": {
            "$ref": "#/components/schemas/AnalysisNarrativeContract"
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingMetadataContract"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "context": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "projectName": {
            "type": "string",
            "nullable": true
          },
          "repository": {
            "type": "string",
            "nullable": true
          },
          "rawContentHash": {
            "type": "string",
            "nullable": true
          },
          "rawContentStored": {
            "type": "boolean"
          },
          "rawContent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MatchedRuleContract": {
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "string",
            "nullable": true
          },
          "targetType": {
            "type": "string",
            "nullable": true
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "matchedConditions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "appliedActions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "appliedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "RootCauseCandidateContract": {
        "type": "object",
        "properties": {
          "fingerprint": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "explanation": {
            "type": "string",
            "nullable": true
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "signals": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "likelyCauses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "suggestedFixes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}